New-events: A change to the structure of the message returned by Event.next
It is now an XMLRPC structure containing two fields - 'events', containing
the list of events as before, and 'valid_ref_counts', containing the number
of objects in each table. Clients should check their caches to ensure that
the counts match. If not, call 'get_all' on the appropriate table to prune
any no-longer-valid entries.

This is required because xapi can't keep track of all objects it has ever
deleted - it tracks up to 500 deleted objects (per table). So under normal
circumstances, the counts should always match up after the client has
applied the diffs. If they don't, that's when the client must check what it
has in its local cache.

Signed-off-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>