Create DB Copy in slave memory
Kick off a thread at the startup of a slave, that talks to the Master's
database at the database level (does not make API level calls). This
slave should maintain a live copy of the database in slave memory. The
changes to the master database should be transferred across as
incremental changes, and applied to the slave's database.

Also plumb in the ability for repointing API level calls to use this
local database.

Signed-off-by: Richard Davies <richard.davies@citrix.com>