Periodically write out slave_db to disk
The slave database is stored in memory, and normally the slave would
have a copy of this in it's fs in a copy from the master. However, it's
more efficient for the slave to just periodically write out its own copy
of the database to the fs itself.

Originally this used `Unix.gettimeofday`, but it was pointed out that
this suffers if the machine's clock is changed. Instead this uses Mtime
to avoid this.

This needs to be expanded by having the writing out to disk vs copy from
master option be configurable and read from xapi.conf

Signed-off-by: Richard Davies <richard.davies@citrix.com>