Change to use fold_left for db changes
Fold_right is not tail recursive, and in the case where we are using it,
the list that we pass in _could_ be large. As such we should use
fold_left.
This means that the 'make_X' functions now must have an explicit db
parameter.

This change also corrects what we do when we've lost some deletes along
the way, to 'do the right thing' and start fresh correctly.

This change also corrects the indentation on writing the slave db to
disk to ensure that it actually only happens when the condition is met.

Signed-off-by: Richard Davies <richard.davies@citrix.com>