Tidying and efficiency
List.rev_map is tail recursive, and should be used in place of List.map
if there's a chance that the list argument will be large. However to
return the result to the original orientation, a wrapping call to
List.rev is required. In a few cases in my code, the ordering doesn't
matter, so the List.rev call can be left off

I've also renamed a function or two to tidy them up (acronyms to full
titles, 'funny' titles to proper ones)

I've also removed parts of functions that weren't required, ie
parameters that can be dropped, or variable assignments that aren't
necessary

I've also removed an if that threw an exception, as I was using this for
debugging purposes, and it shouldn't exist in the code in production

Signed-off-by: Richard Davies <richard.davies@citrix.com>