Use new task handle instead of sessions
For the event from loops, if we are in xapi code, we should not be using
sessions. Instead, we should use something like a task created in the
database that we can alter/delete to cause threads to wakeup, rather
than logging out a session.
This is used wherever we have an event loop, irrespective of whether or
not it originally had a session.
The old path through `from` (with a session) is preserved, as this is
the path that external clients will use

Signed-off-by: Richard Davies <richard.davies@citrix.com>