Don't use session for event from call internally
Even though xapi_vm_clone has access to a session due to it's scope, it
shouldn't make use of it. Instead it should use the task format to grab
control back of the function. In this case, the
'consider_touching_session' task should not be used, but instead one
that considers killing the task. This also uses mtime rather than
Unix.gettimeofday because it's continuous and better

Signed-off-by: Richard Davies <richard.davies@citrix.com>