Don't care about the delete count
We no longer transfer information about the delete list from the master
to the slave, and instead the delete list on the slave only reflects
what has been deleted from that local database. This means that if the
slave misses the entire lifecycle of an object it will know nothing
about that object.

Also we only send out the notifications from a db change if the object
counts are equal. This is because if the object counts are not equal
then the slave database has fallen out of line with the master database,
and as such the slave db has been thrown away to be rebuilt on the next
loop

Signed-off-by: Richard Davies <richard.davies@citrix.com>