Add in logic for positive changes
When making a positive change to the slave db following receiving delta
from the master, there is some logic to be determined as to what apply
functions should be made to the master. This is two-fold, as it's for
both the application to the database, and then afterwards (separate for
mutex control) the notifications of actions happening.

Signed-off-by: Richard Davies <richard.davies@citrix.com>