Created != Modified
When making changes to the slave db, we need to preserve the timestamps
for created and modified from the master. This behaviour is now also
checked in the unit tests.

We now also don't create a session when we query the master from the
slave

And we have better control of the slave db mutex when sending out
notifications

Signed-off-by: Richard Davies <richard.davies@citrix.com>