fixup! Replace event_wait UPDATES.get call with waiting for events on the local DB
The task may not be present in the database yet, do not throw an exception if it is missing
(e.g. because we've just created it and haven't received/processed its creation event yet).

Also fix logic error in procesing event deltas: if the event is not a
task do not assert.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>